daemon: move errToResponse tests out of api_test.go
to either api_snaps_test.go or response_test.go
