+++ daemon/api_test.go
@@ -21,7 +21,6 @@ package daemon
 
 import (
 	"bytes"
-	"errors"
 	"fmt"
 	"mime/multipart"
 	"net/http"
@@ -29,13 +28,8 @@ import (
 
 	"gopkg.in/check.v1"
 
-	"github.com/snapcore/snapd/arch"
 	"github.com/snapcore/snapd/client"
 	"github.com/snapcore/snapd/overlord/auth"
-	"github.com/snapcore/snapd/overlord/snapstate"
-	"github.com/snapcore/snapd/snap"
-	"github.com/snapcore/snapd/snap/channel"
-	"github.com/snapcore/snapd/snap/snaptest"
 	"github.com/snapcore/snapd/store"
 	"github.com/snapcore/snapd/testutil"
 )
@@ -473,247 +467,4 @@ func (s *apiSuite) TestLogsNoServices(c *check.C) {
 	rsp := getLogs(logsCmd, req, nil).(*resp)
 	c.Assert(rsp.Status, check.Equals, 404)
 	c.Assert(rsp.Type, check.Equals, ResponseTypeError)
-}
-
-type fakeNetError struct {
-	message   string
-	timeout   bool
-	temporary bool
-}
-
-func (e fakeNetError) Error() string   { return e.message }
-func (e fakeNetError) Timeout() bool   { return e.timeout }
-func (e fakeNetError) Temporary() bool { return e.temporary }
-
-func (s *apiSuite) TestErrToResponseNoSnapsDoesNotPanic(c *check.C) {
-	si := &snapInstruction{Action: "frobble"}
-	errors := []error{
-		store.ErrSnapNotFound,
-		&store.RevisionNotAvailableError{},
-		store.ErrNoUpdateAvailable,
-		store.ErrLocalSnap,
-		&snap.AlreadyInstalledError{Snap: "foo"},
-		&snap.NotInstalledError{Snap: "foo"},
-		&snapstate.SnapNeedsDevModeError{Snap: "foo"},
-		&snapstate.SnapNeedsClassicError{Snap: "foo"},
-		&snapstate.SnapNeedsClassicSystemError{Snap: "foo"},
-		fakeNetError{message: "other"},
-		fakeNetError{message: "timeout", timeout: true},
-		fakeNetError{message: "temp", temporary: true},
-		errors.New("some other error"),
-	}
-
-	for _, err := range errors {
-		rsp := si.errToResponse(err)
-		com := check.Commentf("%v", err)
-		c.Check(rsp, check.NotNil, com)
-		status := rsp.(*resp).Status
-		c.Check(status/100 == 4 || status/100 == 5, check.Equals, true, com)
-	}
-}
-
-func (s *apiSuite) TestErrToResponseForRevisionNotAvailable(c *check.C) {
-	si := &snapInstruction{Action: "frobble", Snaps: []string{"foo"}}
-
-	thisArch := arch.DpkgArchitecture()
-
-	err := &store.RevisionNotAvailableError{
-		Action:  "install",
-		Channel: "stable",
-		Releases: []channel.Channel{
-			snaptest.MustParseChannel("beta", thisArch),
-		},
-	}
-	rsp := si.errToResponse(err).(*resp)
-	c.Check(rsp, check.DeepEquals, &resp{
-		Status: 404,
-		Type:   ResponseTypeError,
-		Result: &errorResult{
-			Message: "no snap revision on specified channel",
-			Kind:    client.ErrorKindSnapChannelNotAvailable,
-			Value: map[string]interface{}{
-				"snap-name":    "foo",
-				"action":       "install",
-				"channel":      "stable",
-				"architecture": thisArch,
-				"releases": []map[string]interface{}{
-					{"architecture": thisArch, "channel": "beta"},
-				},
-			},
-		},
-	})
-
-	err = &store.RevisionNotAvailableError{
-		Action:  "install",
-		Channel: "stable",
-		Releases: []channel.Channel{
-			snaptest.MustParseChannel("beta", "other-arch"),
-		},
-	}
-	rsp = si.errToResponse(err).(*resp)
-	c.Check(rsp, check.DeepEquals, &resp{
-		Status: 404,
-		Type:   ResponseTypeError,
-		Result: &errorResult{
-			Message: "no snap revision on specified architecture",
-			Kind:    client.ErrorKindSnapArchitectureNotAvailable,
-			Value: map[string]interface{}{
-				"snap-name":    "foo",
-				"action":       "install",
-				"channel":      "stable",
-				"architecture": thisArch,
-				"releases": []map[string]interface{}{
-					{"architecture": "other-arch", "channel": "beta"},
-				},
-			},
-		},
-	})
-
-	err = &store.RevisionNotAvailableError{}
-	rsp = si.errToResponse(err).(*resp)
-	c.Check(rsp, check.DeepEquals, &resp{
-		Status: 404,
-		Type:   ResponseTypeError,
-		Result: &errorResult{
-			Message: "no snap revision available as specified",
-			Kind:    client.ErrorKindSnapRevisionNotAvailable,
-			Value:   "foo",
-		},
-	})
-}
-
-func (s *apiSuite) TestErrToResponseForChangeConflict(c *check.C) {
-	si := &snapInstruction{Action: "frobble", Snaps: []string{"foo"}}
-
-	err := &snapstate.ChangeConflictError{Snap: "foo", ChangeKind: "install"}
-	rsp := si.errToResponse(err).(*resp)
-	c.Check(rsp, check.DeepEquals, &resp{
-		Status: 409,
-		Type:   ResponseTypeError,
-		Result: &errorResult{
-			Message: `snap "foo" has "install" change in progress`,
-			Kind:    client.ErrorKindSnapChangeConflict,
-			Value: map[string]interface{}{
-				"snap-name":   "foo",
-				"change-kind": "install",
-			},
-		},
-	})
-
-	// only snap
-	err = &snapstate.ChangeConflictError{Snap: "foo"}
-	rsp = si.errToResponse(err).(*resp)
-	c.Check(rsp, check.DeepEquals, &resp{
-		Status: 409,
-		Type:   ResponseTypeError,
-		Result: &errorResult{
-			Message: `snap "foo" has changes in progress`,
-			Kind:    client.ErrorKindSnapChangeConflict,
-			Value: map[string]interface{}{
-				"snap-name": "foo",
-			},
-		},
-	})
-
-	// only kind
-	err = &snapstate.ChangeConflictError{Message: "specific error msg", ChangeKind: "some-global-op"}
-	rsp = si.errToResponse(err).(*resp)
-	c.Check(rsp, check.DeepEquals, &resp{
-		Status: 409,
-		Type:   ResponseTypeError,
-		Result: &errorResult{
-			Message: "specific error msg",
-			Kind:    client.ErrorKindSnapChangeConflict,
-			Value: map[string]interface{}{
-				"change-kind": "some-global-op",
-			},
-		},
-	})
-}
-
-func (s *apiSuite) TestErrToResponseInsufficentSpace(c *check.C) {
-	err := &snapstate.InsufficientSpaceError{
-		Snaps:      []string{"foo", "bar"},
-		ChangeKind: "some-change",
-		Path:       "/path",
-		Message:    "specific error msg",
-	}
-	rsp := errToResponse(err, nil, BadRequest, "%s: %v", "ERR").(*resp)
-	c.Check(rsp, check.DeepEquals, &resp{
-		Status: 507,
-		Type:   ResponseTypeError,
-		Result: &errorResult{
-			Message: "specific error msg",
-			Kind:    client.ErrorKindInsufficientDiskSpace,
-			Value: map[string]interface{}{
-				"snap-names":  []string{"foo", "bar"},
-				"change-kind": "some-change",
-			},
-		},
-	})
-}
-
-func (s *apiSuite) TestErrToResponse(c *check.C) {
-	aie := &snap.AlreadyInstalledError{Snap: "foo"}
-	nie := &snap.NotInstalledError{Snap: "foo"}
-	cce := &snapstate.ChangeConflictError{Snap: "foo"}
-	ndme := &snapstate.SnapNeedsDevModeError{Snap: "foo"}
-	nc := &snapstate.SnapNotClassicError{Snap: "foo"}
-	nce := &snapstate.SnapNeedsClassicError{Snap: "foo"}
-	ncse := &snapstate.SnapNeedsClassicSystemError{Snap: "foo"}
-	netoe := fakeNetError{message: "other"}
-	nettoute := fakeNetError{message: "timeout", timeout: true}
-	nettmpe := fakeNetError{message: "temp", temporary: true}
-
-	e := errors.New("other error")
-
-	sa1e := &store.SnapActionError{Refresh: map[string]error{"foo": store.ErrSnapNotFound}}
-	sa2e := &store.SnapActionError{Refresh: map[string]error{
-		"foo": store.ErrSnapNotFound,
-		"bar": store.ErrSnapNotFound,
-	}}
-	saOe := &store.SnapActionError{Other: []error{e}}
-	// this one can't happen (but fun to test):
-	saXe := &store.SnapActionError{Refresh: map[string]error{"foo": sa1e}}
-
-	makeErrorRsp := func(kind client.ErrorKind, err error, value interface{}) Response {
-		return SyncResponse(&resp{
-			Type:   ResponseTypeError,
-			Result: &errorResult{Message: err.Error(), Kind: kind, Value: value},
-			Status: 400,
-		}, nil)
-	}
-
-	tests := []struct {
-		err         error
-		expectedRsp Response
-	}{
-		{store.ErrSnapNotFound, SnapNotFound("foo", store.ErrSnapNotFound)},
-		{store.ErrNoUpdateAvailable, makeErrorRsp(client.ErrorKindSnapNoUpdateAvailable, store.ErrNoUpdateAvailable, "")},
-		{store.ErrLocalSnap, makeErrorRsp(client.ErrorKindSnapLocal, store.ErrLocalSnap, "")},
-		{aie, makeErrorRsp(client.ErrorKindSnapAlreadyInstalled, aie, "foo")},
-		{nie, makeErrorRsp(client.ErrorKindSnapNotInstalled, nie, "foo")},
-		{ndme, makeErrorRsp(client.ErrorKindSnapNeedsDevMode, ndme, "foo")},
-		{nc, makeErrorRsp(client.ErrorKindSnapNotClassic, nc, "foo")},
-		{nce, makeErrorRsp(client.ErrorKindSnapNeedsClassic, nce, "foo")},
-		{ncse, makeErrorRsp(client.ErrorKindSnapNeedsClassicSystem, ncse, "foo")},
-		{cce, SnapChangeConflict(cce)},
-		{nettoute, makeErrorRsp(client.ErrorKindNetworkTimeout, nettoute, "")},
-		{netoe, BadRequest("ERR: %v", netoe)},
-		{nettmpe, BadRequest("ERR: %v", nettmpe)},
-		{e, BadRequest("ERR: %v", e)},
-
-		// action error unwrapping:
-		{sa1e, SnapNotFound("foo", store.ErrSnapNotFound)},
-		{saXe, SnapNotFound("foo", store.ErrSnapNotFound)},
-		// action errors, unwrapped:
-		{sa2e, BadRequest(`ERR: cannot refresh: snap not found: "bar", "foo"`)},
-		{saOe, BadRequest("ERR: cannot refresh, install, or download: other error")},
-	}
-
-	for _, t := range tests {
-		com := check.Commentf("%v", t.err)
-		rsp := errToResponse(t.err, []string{"foo"}, BadRequest, "%s: %v", "ERR")
-		c.Check(rsp, check.DeepEquals, t.expectedRsp, com)
-	}
 }
