+++ daemon/response_test.go
@@ -21,6 +21,7 @@ package daemon_test
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io/ioutil"
 	"net/http"
@@ -30,7 +31,11 @@ import (
 
 	"gopkg.in/check.v1"
 
+	"github.com/snapcore/snapd/client"
 	"github.com/snapcore/snapd/daemon"
+	"github.com/snapcore/snapd/overlord/snapstate"
+	"github.com/snapcore/snapd/snap"
+	"github.com/snapcore/snapd/store"
 )
 
 type responseSuite struct{}
@@ -138,4 +143,101 @@ func (responseSuite) TestErrorResponderDoesNotPrintfAlways(c *check.C) {
 	c.Assert(json.NewDecoder(rec.Body).Decode(&v), check.IsNil)
 
 	c.Check(v.Result.Message, check.Equals, "system memory below 1%.")
+}
+
+type fakeNetError struct {
+	message   string
+	timeout   bool
+	temporary bool
+}
+
+func (e fakeNetError) Error() string   { return e.message }
+func (e fakeNetError) Timeout() bool   { return e.timeout }
+func (e fakeNetError) Temporary() bool { return e.temporary }
+
+func (s *responseSuite) TestErrToResponse(c *check.C) {
+	aie := &snap.AlreadyInstalledError{Snap: "foo"}
+	nie := &snap.NotInstalledError{Snap: "foo"}
+	cce := &snapstate.ChangeConflictError{Snap: "foo"}
+	ndme := &snapstate.SnapNeedsDevModeError{Snap: "foo"}
+	nc := &snapstate.SnapNotClassicError{Snap: "foo"}
+	nce := &snapstate.SnapNeedsClassicError{Snap: "foo"}
+	ncse := &snapstate.SnapNeedsClassicSystemError{Snap: "foo"}
+	netoe := fakeNetError{message: "other"}
+	nettoute := fakeNetError{message: "timeout", timeout: true}
+	nettmpe := fakeNetError{message: "temp", temporary: true}
+
+	e := errors.New("other error")
+
+	sa1e := &store.SnapActionError{Refresh: map[string]error{"foo": store.ErrSnapNotFound}}
+	sa2e := &store.SnapActionError{Refresh: map[string]error{
+		"foo": store.ErrSnapNotFound,
+		"bar": store.ErrSnapNotFound,
+	}}
+	saOe := &store.SnapActionError{Other: []error{e}}
+	// this one can't happen (but fun to test):
+	saXe := &store.SnapActionError{Refresh: map[string]error{"foo": sa1e}}
+
+	makeErrorRsp := func(kind client.ErrorKind, err error, value interface{}) daemon.Response {
+		return daemon.SyncResponse(&daemon.Resp{
+			Type:   daemon.ResponseTypeError,
+			Result: &daemon.ErrorResult{Message: err.Error(), Kind: kind, Value: value},
+			Status: 400,
+		}, nil)
+	}
+
+	tests := []struct {
+		err         error
+		expectedRsp daemon.Response
+	}{
+		{store.ErrSnapNotFound, daemon.SnapNotFound("foo", store.ErrSnapNotFound)},
+		{store.ErrNoUpdateAvailable, makeErrorRsp(client.ErrorKindSnapNoUpdateAvailable, store.ErrNoUpdateAvailable, "")},
+		{store.ErrLocalSnap, makeErrorRsp(client.ErrorKindSnapLocal, store.ErrLocalSnap, "")},
+		{aie, makeErrorRsp(client.ErrorKindSnapAlreadyInstalled, aie, "foo")},
+		{nie, makeErrorRsp(client.ErrorKindSnapNotInstalled, nie, "foo")},
+		{ndme, makeErrorRsp(client.ErrorKindSnapNeedsDevMode, ndme, "foo")},
+		{nc, makeErrorRsp(client.ErrorKindSnapNotClassic, nc, "foo")},
+		{nce, makeErrorRsp(client.ErrorKindSnapNeedsClassic, nce, "foo")},
+		{ncse, makeErrorRsp(client.ErrorKindSnapNeedsClassicSystem, ncse, "foo")},
+		{cce, daemon.SnapChangeConflict(cce)},
+		{nettoute, makeErrorRsp(client.ErrorKindNetworkTimeout, nettoute, "")},
+		{netoe, daemon.BadRequest("ERR: %v", netoe)},
+		{nettmpe, daemon.BadRequest("ERR: %v", nettmpe)},
+		{e, daemon.BadRequest("ERR: %v", e)},
+
+		// action error unwrapping:
+		{sa1e, daemon.SnapNotFound("foo", store.ErrSnapNotFound)},
+		{saXe, daemon.SnapNotFound("foo", store.ErrSnapNotFound)},
+		// action errors, unwrapped:
+		{sa2e, daemon.BadRequest(`ERR: cannot refresh: snap not found: "bar", "foo"`)},
+		{saOe, daemon.BadRequest("ERR: cannot refresh, install, or download: other error")},
+	}
+
+	for _, t := range tests {
+		com := check.Commentf("%v", t.err)
+		rsp := daemon.ErrToResponse(t.err, []string{"foo"}, daemon.BadRequest, "%s: %v", "ERR")
+		c.Check(rsp, check.DeepEquals, t.expectedRsp, com)
+	}
+}
+
+func (s *responseSuite) TestErrToResponseInsufficentSpace(c *check.C) {
+	err := &snapstate.InsufficientSpaceError{
+		Snaps:      []string{"foo", "bar"},
+		ChangeKind: "some-change",
+		Path:       "/path",
+		Message:    "specific error msg",
+	}
+	rsp := daemon.ErrToResponse(err, nil, daemon.BadRequest, "%s: %v", "ERR").(*daemon.Resp)
+	c.Check(rsp, check.DeepEquals, &daemon.Resp{
+		Status: 507,
+		Type:   daemon.ResponseTypeError,
+		Result: &daemon.ErrorResult{
+			Message: "specific error msg",
+			Kind:    client.ErrorKindInsufficientDiskSpace,
+			Value: map[string]interface{}{
+				"snap-names":  []string{"foo", "bar"},
+				"change-kind": "some-change",
+			},
+		},
+	})
 }
